Avoid nuget HTTP cache when installing/updating
The nuget cache is used heavily when installing or updating tools. It's not unusual for a recently published package (which does show up in our explicit check against the feed) to fail to install the latest version if a locally cached version exists too.

Since we check for updates against the actual remote feed, we know when there's an update, and the cache can get in the way of consistency here. We therefore pass `--no-cache` when running install/update commands.

Fixes #6